+++ src/renderer/components/charts/BarChart.js
