+++ src/renderer/components/charts/CountsWidget.js
