+++ src/renderer/components/charts/LineChart.js
