+++ src/renderer/components/charts/PieChart.js
