+++ src/renderer/components/index.js
@@ -1,13 +1,13 @@
-export { default as CountsWidget } from './CountsWidget';
-export { default as BarChart } from './BarChart';
+export { default as CountsWidget } from './charts/CountsWidget';
+export { default as BarChart } from './charts/BarChart';
 export { default as Header } from './Header';
-export { default as InterviewWidget } from './InterviewWidget';
-export { default as LineChart } from './LineChart';
+export { default as InterviewWidget } from './charts/InterviewWidget';
+export { default as LineChart } from './charts/LineChart';
 export { default as MenuItem } from './MenuItem';
 export { default as Modal } from './Modal';
 export { default as PanelItem } from './PanelItem';
 export { default as PairPrompt } from './pairing/PairPrompt';
 export { default as PairPin } from './pairing/PairPin';
-export { default as PieChart } from './PieChart';
+export { default as PieChart } from './charts/PieChart';
 export { default as ServerPanel } from './ServerPanel';
 export { default as TabBar } from './TabBar';
