+++ src/renderer/styles/main.scss
@@ -1,4 +1,4 @@
-@import '~network-canvas-ui/src/styles/all';
+@import '~network-canvas-ui/lib/styles/all';
 @import 'variables';
 @import 'helpers/all';
 @import 'ui/all';
